Duane Analysis

This function performs a Duane analysis (1962) doi:10.1109/TA.1964.4319640 on failure data by fitting a log-log linear regression of cumulative MTBF versus cumulative time.

Usage

duane(times, failures, conf.int = FALSE, conf.level = 0.95)

Arguments

times

A numeric vector of cumulative failure times.

failures

A numeric vector of the number of failures at each corresponding time in times.

conf.int

Logical; whether to compute confidence bounds (default: FALSE).

conf.level

Confidence level for the confidence bounds (default: 0.95).

Value

A list of class "duane" containing:

model

The fitted lm object.

logLik

The log-likelihood of the fitted model.

AIC

Akaike Information Criterion.

BIC

Bayesian Information Criterion.

conf.level

The confidence level, if applicable.

Cumulative_Time

The cumulative operating times.

Cumulative_MTBF

The cumulative mean time between failures.

Fitted_Values

The fitted values on the MTBF scale.

Confidence_Bounds

Matrix of fitted values and confidence bounds on the MTBF scale (if applicable).

See also

Examples

times <- c(100, 200, 300, 400, 500)
failures <- c(1, 2, 1, 3, 2)
fit <- duane(times, failures, conf.int = TRUE, conf.level = 0.90)
print(fit)
#> Duane Analysis Result
#> ----------------------
#> Linear model (log-log scale): log(MTBF) ~ log(Time)
#> 
#> Coefficients:
#>                Estimate Std. Error
#> (Intercept)   3.6144974 0.35199619
#> log_cum_times 0.2013244 0.05624037
#> 
#> Log-likelihood: 4.78
#> AIC: -3.55, BIC: -4.72
#> Confidence level: 90.0%