Reliability Growth Analysis.

This function performs reliability growth analysis using the Crow-AMSAA model by Crow (1975) https://apps.dtic.mil/sti/citations/ADA020296 or piecewise Weibull NHPP model by Guo et al. (2010) doi:10.1109/RAMS.2010.5448029.

Usage

rga(
  times,
  failures,
  model_type = "Crow-AMSAA",
  breakpoints = NULL,
  conf_level = 0.95
)

Arguments

times

A vector of cumulative times at which failures occurred.

failures

A vector of the number of failures at each corresponding time in times.

model_type

The model type. Either Crow-AMSAA (default) or Piecewise Weibull NHPP with change point detection.

breakpoints

An optional vector of breakpoints for the Piecewise Weibull NHPP model.

conf_level

The desired confidence level, which defaults to 95%.

Value

The function returns an object of class rga that contains the results for the model.

Examples

times <- c(100, 200, 300, 400, 500)
failures <- c(1, 2, 1, 3, 2)
result <- rga(times, failures)
print(result)
#> Reliability Growth Analysis (RGA)
#> ---------------------------------
#> Model Type: Crow-AMSAA 
#> 
#> Weibull Parameters (per segment):
#>   Beta: 0.7987
#>    Beta: 0.0562
#>    Beta: 14.2011
#>    Beta: 0.0008
#>   Lambda: -3.6145
#>    Lambda: 0.3520
#>    Lambda: -10.2686
#>    Lambda: 0.0020
#> 
#> Goodness of Fit:
#>   AIC: -3.55
#>   BIC: -4.72